refactor trading page to use callback
