+++ src/javascript/binary/websocket_pages/mb_trade/mb_tradepage.js
@@ -7,7 +7,7 @@ const MBPrice              = require('./mb_price');
 const MBProcess            = require('./mb_process');
 const MBSymbols            = require('./mb_symbols');
 const TradingAnalysis      = require('../trade/analysis');
-const chartFrameCleanup    = require('../trade/common').chartFrameCleanup;
+const chartFrameCleanup    = require('../trade/charts/chart_frame').chartFrameCleanup;
 const localize             = require('../../base/localize').localize;
 const State                = require('../../base/storage').State;
 const JapanPortfolio       = require('../../../binary_japan/trade_japan/portfolio');
