+++ src/javascript/binary/websocket_pages/trade/beta/event.js
@@ -5,7 +5,9 @@ const Contract_Beta                  = require('./contract');
 const Durations_Beta                 = require('./duration');
 const Price_Beta                     = require('./price');
 const Process_Beta                   = require('./process');
+const chartFrameSource               = require('../charts/chart_frame').chartFrameSource;
 const Defaults                       = require('../defaults');
+const GetTicks                       = require('../get_ticks');
 const setFormPlaceholderContent_Beta = require('../set_values').setFormPlaceholderContent_Beta;
 const Tick                           = require('../tick');
 const commonTrading                  = require('../common');
@@ -41,8 +43,8 @@ const TradingEvents_Beta = (() => {
             // it will default to proper one
             Defaults.remove('formname');
             Defaults.remove('underlying');
-            Process_Beta.processMarket_Beta(1);
-            commonTrading.chartFrameSource();
+            Process_Beta.processMarket_Beta();
+            chartFrameSource();
         };
 
         const market_nav_element = document.getElementById('contract_markets');
@@ -91,7 +93,7 @@ const TradingEvents_Beta = (() => {
         if (underlying_element) {
             underlying_element.addEventListener('change', (e) => {
                 if (e.target) {
-                    commonTrading.chartFrameSource();
+                    chartFrameSource();
                     commonTrading.showFormOverlay();
                     commonTrading.showPriceOverlay();
                     if (e.target.selectedIndex < 0) {
@@ -111,7 +113,7 @@ const TradingEvents_Beta = (() => {
                     // forget the old tick id i.e. close the old tick stream
                     Process_Beta.processForgetTicks_Beta();
                     // get ticks for current underlying
-                    Tick.request(underlying);
+                    GetTicks.request(underlying);
                     commonTrading.displayTooltip_Beta(Defaults.get('market'), underlying);
                 }
             });
