+++ src/javascript/binary/websocket_pages/trade/beta/tradepage.js
@@ -4,10 +4,10 @@ const Message_Beta              = require('./message');
 const Price_Beta                = require('./price');
 const forgetTradingStreams_Beta = require('./process').forgetTradingStreams_Beta;
 const commonTrading             = require('../common');
+const chartFrameCleanup         = require('../charts/chart_frame').chartFrameCleanup;
 const displayCurrencies         = require('../currency');
 const Defaults                  = require('../defaults');
-const Notifications             = require('../notifications');
-const Symbols                   = require('../symbols');
+const processActiveSymbols      = require('../process').processActiveSymbols;
 const PortfolioInit             = require('../../user/account/portfolio/portfolio.init');
 const ViewPopup                 = require('../../user/view_popup/view_popup');
 const BinaryPjax                = require('../../../base/binary_pjax');
@@ -33,9 +33,6 @@ const TradePage_Beta = (() => {
             onmessage: (msg) => {
                 Message_Beta.process(msg);
             },
-            onopen: () => {
-                Notifications.hide('CONNECTION_ERROR');
-            },
         });
         Price_Beta.clearFormId();
         if (events_initialized === 0) {
@@ -45,7 +42,7 @@ const TradePage_Beta = (() => {
 
         BinarySocket.send({ payout_currencies: 1 }).then(() => {
             displayCurrencies();
-            Symbols.getSymbols(1);
+            processActiveSymbols();
         });
 
         if (document.getElementById('websocket_form')) {
@@ -181,13 +178,13 @@ const TradePage_Beta = (() => {
         BinarySocket.clear();
         Defaults.clear();
         PortfolioInit.onUnload();
-        commonTrading.chartFrameCleanup();
+        chartFrameCleanup();
     };
 
     const onDisconnect = () => {
         commonTrading.showPriceOverlay();
         commonTrading.showFormOverlay();
-        commonTrading.chartFrameCleanup();
+        chartFrameCleanup();
         commonTrading.clean();
         onLoad();
     };
