+++ src/javascript/binary/websocket_pages/trade/common.js
@@ -4,11 +4,9 @@ const Notifications      = require('./notifications');
 const Symbols            = require('./symbols');
 const Tick               = require('./tick');
 const Client             = require('../../base/client');
-const getLanguage        = require('../../base/language').get;
 const localize           = require('../../base/localize').localize;
 const urlFor             = require('../../base/url').urlFor;
 const isEmptyObject      = require('../../base/utility').isEmptyObject;
-const jpClient           = require('../../common_functions/country_base').jpClient;
 const formatMoney        = require('../../common_functions/currency_to_symbol').formatMoney;
 const addComma           = require('../../common_functions/string_util').addComma;
 const toISOFormat        = require('../../common_functions/string_util').toISOFormat;
@@ -353,21 +351,6 @@ const commonTrading = (() => {
         }
     };
 
-    /*
-     * toggle active class of menu
-     */
-    const toggleActiveNavMenuElement = (nav, event_element) => {
-        const li_elements = nav.getElementsByTagName('li');
-        const classes = event_element.classList;
-
-        if (!classes.contains('active')) {
-            for (let i = 0, len = li_elements.length; i < len; i++) {
-                li_elements[i].classList.remove('active');
-            }
-            classes.add('active');
-        }
-    };
-
     const toggleActiveCatMenuElement = (nav, event_element_id) => {
         const event_element = document.getElementById(event_element_id);
         const li_elements = nav.querySelectorAll('.active, .a-active');
@@ -608,26 +591,6 @@ const commonTrading = (() => {
         return false;
     };
 
-    const updatePurchaseStatus = (final_price, pnl, contract_status) => {
-        $('#contract_purchase_heading').text(localize(contract_status));
-        const $payout = $('#contract_purchase_payout');
-        const $cost   = $('#contract_purchase_cost');
-        const $profit = $('#contract_purchase_profit');
-
-        $payout.html($('<div/>', { text: localize('Buy price') }).append($('<p/>', { text: addComma(Math.abs(pnl)) })));
-        $cost.html($('<div/>', { text: localize('Final price') }).append($('<p/>', { text: addComma(final_price) })));
-        if (!final_price) {
-            $profit.html($('<div/>', { text: localize('Loss') }).append($('<p/>', { text: addComma(pnl) })));
-        } else {
-            $profit.html($('<div/>', { text: localize('Profit') }).append($('<p/>', { text: addComma(Math.round((final_price - pnl) * 100) / 100) })));
-            updateContractBalance(Client.get('balance'));
-        }
-    };
-
-    const updateContractBalance = (balance) => {
-        $('#contract_purchase_balance').text(`${localize('Account balance:')} ${formatMoney(Client.get('currency'), balance)}`);
-    };
-
     const chart_config = {
         type              : 'line',
         lineColor         : '#606060',
@@ -660,41 +623,6 @@ const commonTrading = (() => {
             'date_start', 'expiry_type', 'expiry_date', 'expirt_time', 'duration_units', 'diration_value',
             'amount', 'amount_type', 'currency', 'stop_loss', 'stop_type', 'stop_profit', 'amount_per_point', 'prediction');
         location.reload();
-    };
-
-    const showHighchart = () => {
-        if (window.chartAllowed) {
-            chartFrameSource();
-        } else {
-            chartFrameCleanup();
-            $('#trade_live_chart').hide();
-            $('#chart-error').text(localize('Chart is not available for this underlying.')).show();
-        }
-    };
-
-    const chartFrameCleanup = () => {
-        /*
-         * Prevent IE memory leak (http://stackoverflow.com/questions/8407946).
-         */
-        const chart_frame = document.getElementById('chart_frame');
-        if (chart_frame) {
-            chart_frame.src = 'about:blank';
-        }
-    };
-
-    const chartFrameSource = () => {
-        if ($('#tab_graph').hasClass('active') && (sessionStorage.getItem('old_underlying') !== sessionStorage.getItem('underlying') || /^(|about:blank)$/.test($('#chart_frame').attr('src')))) {
-            chartFrameCleanup();
-            setChartSource();
-            sessionStorage.setItem('old_underlying', document.getElementById('underlying').value);
-        }
-        $('#chart-error').hide();
-        $('#trade_live_chart').show();
-    };
-
-    const setChartSource = () => {
-        const is_ja = !!jpClient();
-        document.getElementById('chart_frame').src = `https://webtrader.binary.com?affiliates=true&instrument=${document.getElementById('underlying').value}&timePeriod=1t&gtm=true&lang=${getLanguage().toLowerCase()}&hideOverlay=${is_ja}&hideShare=${is_ja}&timezone=GMT+${(is_ja ? '9' : '0')}&hideFooter=${is_ja}`;
     };
 
     // ============= Functions used in /trading_beta =============
@@ -816,7 +744,6 @@ const commonTrading = (() => {
         hideOverlayContainer           : hideOverlayContainer,
         getContractCategoryTree        : getContractCategoryTree,
         resetPriceMovement             : resetPriceMovement,
-        toggleActiveNavMenuElement     : toggleActiveNavMenuElement,
         toggleActiveCatMenuElement     : toggleActiveCatMenuElement,
         displayCommentPrice            : displayCommentPrice,
         displayCommentSpreads          : displayCommentSpreads,
@@ -827,13 +754,8 @@ const commonTrading = (() => {
         durationOrder                  : duration => duration_order[duration],
         displayTooltip                 : displayTooltip,
         selectOption                   : selectOption,
-        updatePurchaseStatus           : updatePurchaseStatus,
-        updateContractBalance          : updateContractBalance,
         updateWarmChart                : updateWarmChart,
         reloadPage                     : reloadPage,
-        showHighchart                  : showHighchart,
-        chartFrameCleanup              : chartFrameCleanup,
-        chartFrameSource               : chartFrameSource,
         displayContractForms           : displayContractForms,
         displayMarkets                 : displayMarkets,
         toggleActiveNavMenuElement_Beta: toggleActiveNavMenuElement_Beta,
