+++ src/javascript/binary/websocket_pages/trade/event.js
@@ -2,15 +2,19 @@ const moment                     = require('moment');
 const TradingAnalysis            = require('./analysis');
 const Barriers                   = require('./barriers');
 const commonTrading              = require('./common');
-const Contract                   = require('./contract');
+const chartFrameSource           = require('./charts/chart_frame').chartFrameSource;
 const Defaults                   = require('./defaults');
 const Durations                  = require('./duration');
+const GetTicks                   = require('./get_ticks');
+const Notifications              = require('./notifications');
 const Price                      = require('./price');
 const Process                    = require('./process');
+const Purchase                   = require('./purchase');
 const setFormPlaceholderContent  = require('./set_values').setFormPlaceholderContent;
 const getStartDateNode           = require('./common_independent').getStartDateNode;
 const Tick                       = require('./tick');
 const BinaryPjax                 = require('../../base/binary_pjax');
+const GTM                        = require('../../base/gtm');
 const dateValueChanged           = require('../../common_functions/common_functions').dateValueChanged;
 const elementTextContent         = require('../../common_functions/common_functions').elementTextContent;
 const isVisible                  = require('../../common_functions/common_functions').isVisible;
@@ -41,8 +45,8 @@ const TradingEvents = (() => {
             // it will default to proper one
             Defaults.remove('formname');
             Defaults.remove('underlying');
-            Process.processMarket(1);
-            commonTrading.chartFrameSource();
+            Process.processMarket();
+            chartFrameSource();
         };
 
         const market_nav_element = document.getElementById('contract_markets');
@@ -91,7 +95,7 @@ const TradingEvents = (() => {
         if (underlying_element) {
             underlying_element.addEventListener('change', (e) => {
                 if (e.target) {
-                    commonTrading.chartFrameSource();
+                    chartFrameSource();
                     commonTrading.showFormOverlay();
                     commonTrading.showPriceOverlay();
                     if (e.target.selectedIndex < 0) {
@@ -106,16 +110,24 @@ const TradingEvents = (() => {
 
                     commonTrading.updateWarmChart();
 
-                    Contract.getContracts(underlying);
+                    getContracts(underlying);
 
                     // forget the old tick id i.e. close the old tick stream
                     Process.processForgetTicks();
                     // get ticks for current underlying
-                    Tick.request(underlying);
+                    GetTicks.request(underlying);
                     commonTrading.displayTooltip(Defaults.get('market'), underlying);
                 }
             });
         }
+
+        const getContracts = (underlying) => {
+            BinarySocket.send({ contracts_for: underlying }).then((response) => {
+                Notifications.hide('CONNECTION_ERROR');
+                Process.processContract(response);
+                window.contracts_for = response;
+            });
+        };
 
         /*
          * bind event to change in duration amount, request new price
@@ -322,7 +334,10 @@ const TradingEvents = (() => {
                 }, this);
                 if (id && ask_price) {
                     $('.purchase_button').css('visibility', 'hidden');
-                    BinarySocket.send(params);
+                    BinarySocket.send(params).then((response) => {
+                        Purchase.display(response);
+                        GTM.pushPurchaseData(response);
+                    });
                     Price.incrFormId();
                     Price.processForgetProposals();
                 }
