+++ src/javascript/binary/websocket_pages/trade/purchase.js
@@ -1,8 +1,8 @@
-const commonTrading      = require('./common');
 const Contract           = require('./contract');
 const Symbols            = require('./symbols');
 const Tick               = require('./tick');
 const TickDisplay        = require('./tick_trade');
+const updateValues       = require('./update_values');
 const localize           = require('../../base/localize').localize;
 const elementInnerHtml   = require('../../common_functions/common_functions').elementInnerHtml;
 const elementTextContent = require('../../common_functions/common_functions').elementTextContent;
@@ -85,7 +85,7 @@ const Purchase = (() => {
                 elementInnerHtml(profit, `${localize('Potential Profit')} <p>${profit_value}</p>`);
             }
 
-            commonTrading.updateContractBalance(receipt.balance_after);
+            updateValues.updateContractBalance(receipt.balance_after);
 
             if (show_chart) {
                 chart.show();
@@ -235,7 +235,7 @@ const Purchase = (() => {
                         contract_status = localize('This contract lost');
                     }
 
-                    commonTrading.updatePurchaseStatus(final_price, pnl, contract_status);
+                    updateValues.updatePurchaseStatus(final_price, pnl, contract_status);
                 }
 
                 duration--;
