+++ src/javascript/binary/websocket_pages/trade/tick_trade.js
@@ -1,7 +1,7 @@
 const Highcharts           = require('highcharts');
 const moment               = require('moment');
-const updatePurchaseStatus = require('./common').updatePurchaseStatus;
 const Tick                 = require('./tick');
+const updatePurchaseStatus = require('./update_values').updatePurchaseStatus;
 const ViewPopupUI          = require('../user/view_popup/view_popup.ui');
 const localize             = require('../../base/localize').localize;
 const elementInnerHtml     = require('../../common_functions/common_functions').elementInnerHtml;
